new commit in 2024/11/20-1

+++ 4_DL/day_40/40_1_StartTest.py
@@ -5,6 +5,6 @@ if __name__ == '__main__':
     # x = torch.arange(12, dtype=torch.float32)
     # print(x)
     # print(x.numel())
-    y = torch.arange(24, dtype=torch.float64)
+    y = torch.arange(24, dtype=torch.float)
     print(y)
     print(y.numel())
